add custom `highs` value for `torch.randint`
