+++ src/ml-adder/main.py
@@ -8,39 +8,9 @@ import model as m
 import torch.nn as nn
 import torch.optim as optim
 
+from tqdm import trange
 
-def valid_tensor_shape(t: torch.Tensor):
-    if len(t.shape) != 3:
-        print(
-            'Invalid tensor shape. The tensor must be of size [batch_size, N, M]')
-        print(f'Actual tensor shape: {t.shape}')
-        exit(1)
-
-
-def sum_tensor_by_columns(t: torch.Tensor):
-    valid_tensor_shape(t)
-    a = t[:, :, 0]
-    b = t[:, :, 1]
-    return (a + b).view(t.shape[0], -1, 1)
-
-
-def generate_tensor(n_samples: int, highs: int, lows: int = 0):
-    dim_t = (n_samples, 1, 2)
-    t = torch.randint(lows, highs, dim_t, dtype=torch.float)
-    return t
-
-
-def generate_input_data(n_samples: int):
-    t = generate_tensor(n_samples, 10)
-    st = sum_tensor_by_columns(t)
-    return torch.cat((t, st), dim=-1)
-
-
-def generate_features_and_labels(t: torch.tensor):
-    valid_tensor_shape(t)
-    features = t[:, :, :2]
-    labels = t[:, :, 2].view(t.shape[0], -1, 1)
-    return features, labels
+import data
 
 
 def train(model: nn.Module, loss_fn, dataloader: DataLoader):
@@ -94,22 +64,11 @@ def eval(model_path: str, loss_fn, eval_dataloader: DataLoader):
 
 
 if __name__ == "__main__":
-    batch_size = 32
-    n_samples = 100000
+    batch_size = 64
+    n_samples = 150000
 
-    train_data = generate_input_data(n_samples)
-    train_features, train_labels = generate_features_and_labels(train_data)
-
-    test_data = generate_input_data(1000)
-    test_features, test_labels = generate_features_and_labels(test_data)
-
-    train_set = TensorDataset(train_features, train_labels)
-    test_set = TensorDataset(test_features, test_labels)
-
-    train_loader = DataLoader(train_set, batch_size=batch_size, shuffle=False)
-    test_loader = DataLoader(test_set, batch_size=1, shuffle=False)
-
-    test_loader = DataLoader(test_set)
+    train_loader = data.generate_train_data(n_samples, batch_size)
+    test_loader = data.generate_test_data(5000)
 
     config = m.AdnetConfig(
         input_size=2,
